+++ HISTORY.rst
@@ -34,4 +34,5 @@ History
 * BODY-49: Integrate fix from briefy.ws. (rudaporto)
 * LEICA-46: Update user_id data on all fields to user info map when object is serialized. (rudaporto)
 * Integrate briefy.ws fixes for workflow endpoint POST with empty message attribute on body. (rudaporto)
+* BODY-52: (hotfix) Quote filename for thumbor image signature. (ericof)