+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 * Integrate new fixes on briefy.ws. (rudaporto)
 * Deploy to update briefy.ws. (rudaporto)
 * LEICA-56: New service to return delivery info for a job. (rudaporto)
+* LEICA-58: Update Knack on job approval and rejection. (ericof)
 
 1.0.0
 -----